+++ src/main/java/org/datadog/jenkins/plugins/datadog/apm/JavaConfigurator.java
@@ -4,6 +4,11 @@ import hudson.FilePath;
 import hudson.model.Node;
 import hudson.model.TaskListener;
 import hudson.util.Secret;
+import jenkins.model.Jenkins;
+import org.datadog.jenkins.plugins.datadog.DatadogUtilities;
+import org.datadog.jenkins.plugins.datadog.apm.signature.SignatureVerifier;
+import org.datadog.jenkins.plugins.datadog.clients.HttpClient;
+
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -17,25 +22,19 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import jenkins.model.Jenkins;
-import org.datadog.jenkins.plugins.datadog.DatadogUtilities;
-import org.datadog.jenkins.plugins.datadog.apm.signature.SignatureVerifier;
-import org.datadog.jenkins.plugins.datadog.clients.HttpClient;
 
 final class JavaConfigurator implements TracerConfigurator {
 
     private static final Logger LOGGER = Logger.getLogger(JavaConfigurator.class.getName());
 
     private static final String TRACER_DISTRIBUTION_URL_ENV_VAR = "DATADOG_JENKINS_PLUGIN_TRACER_DISTRIBUTION_URL";
+    private static final String DATADOG_PUBLIC_KEY_ENV_VAR = "DATADOG_JENKINS_PLUGIN_DATADOG_PUBLIC_KEY";
     private static final String DEFAULT_TRACER_DISTRIBUTION_URL = "https://dtdg.co/latest-java-tracer";
     private static final String TRACER_FILE_NAME = "dd-java-agent.jar";
     private static final String TRACER_IGNORE_JENKINS_PROXY_ENV_VAR = "DATADOG_JENKINS_PLUGIN_TRACER_IGNORE_JENKINS_PROXY";
     private static final String TRACER_JAR_CACHE_TTL_ENV_VAR = "DATADOG_JENKINS_PLUGIN_TRACER_JAR_CACHE_TTL_MINUTES";
     private static final int DEFAULT_TRACER_JAR_CACHE_TTL_MINUTES = 60 * 12;
     private static final int TRACER_DOWNLOAD_TIMEOUT_MILLIS = 60_000;
-    private static final String DATADOG_DISTRIBUTION_HOST = "dtdg.co";
-    private static final String MAVEN_CENTRAL_HOST = "repo1.maven.org";
-    private static final String DATADOG_AGENT_MAVEN_DISTRIBUTION_PATH = "/maven2/com/datadoghq/dd-java-agent/";
 
     private final HttpClient httpClient = new HttpClient(TRACER_DOWNLOAD_TIMEOUT_MILLIS);
 
@@ -101,7 +100,7 @@ final class JavaConfigurator implements TracerConfigurator {
     }
 
     private byte[] getTracerSignaturePublicKey(DatadogTracerJobProperty<?> tracerConfig) {
-        return getSetting(tracerConfig, TRACER_DISTRIBUTION_URL_ENV_VAR, SignatureVerifier.DATADOG_PUBLIC_KEY.getBytes(StandardCharsets.UTF_8), String::getBytes);
+        return getSetting(tracerConfig, DATADOG_PUBLIC_KEY_ENV_VAR, SignatureVerifier.DATADOG_PUBLIC_KEY.getBytes(StandardCharsets.UTF_8), String::getBytes);
     }
 
     private <T> T getSetting(DatadogTracerJobProperty<?> tracerConfig, String envVariableName, T defaultValue, Function<String, T> parser) {
@@ -129,27 +128,12 @@ final class JavaConfigurator implements TracerConfigurator {
     }
 
     private String validateUserSuppliedTracerUrl(String distributionUrl) {
-        URL url;
         try {
-            url = new URL(distributionUrl);
+            new URL(distributionUrl);
+            return distributionUrl;
         } catch (MalformedURLException e) {
             throw new IllegalArgumentException("Error while parsing tracer distribution URL: " + distributionUrl, e);
         }
-
-        String host = url.getHost();
-        if (DATADOG_DISTRIBUTION_HOST.equals(host)) {
-            return distributionUrl;
-        }
-
-        if (!MAVEN_CENTRAL_HOST.equals(host)) {
-            throw new IllegalArgumentException("Illegal tracer distribution host: " + host + " (" + distributionUrl + ")");
-        }
-
-        String path = url.getPath();
-        if (!path.startsWith(DATADOG_AGENT_MAVEN_DISTRIBUTION_PATH)) {
-            throw new IllegalArgumentException("Illegal tracer distribution path: " + path + " (" + distributionUrl + ")");
-        }
-        return distributionUrl;
     }
 
     private static Map<String, String> getEnvVariables(DatadogTracerJobProperty<?> tracerConfig,