+++ src/main/java/org/datadog/jenkins/plugins/datadog/flare/DatadogConfigFlare.java
@@ -1,0 +1,42 @@
+package org.datadog.jenkins.plugins.datadog.flare;
+
+import hudson.Extension;
+import hudson.util.XStream2;
+import org.datadog.jenkins.plugins.datadog.DatadogGlobalConfiguration;
+import org.datadog.jenkins.plugins.datadog.DatadogUtilities;
+
+import java.io.IOException;
+import java.io.OutputStream;
+
+@Extension
+public class DatadogConfigFlare implements FlareContributor {
+
+    // TODO use XSTREAM from DatadogGlobalConfiguration following configuration refactor
+    private static final XStream2 XSTREAM;
+
+    static {
+        XSTREAM = new XStream2();
+        XSTREAM.autodetectAnnotations(true);
+    }
+
+    @Override
+    public int order() {
+        return 1;
+    }
+
+    @Override
+    public String getDescription() {
+        return "Plugin configuration";
+    }
+
+    @Override
+    public String getFilename() {
+        return "DatadogGlobalConfiguration.xml";
+    }
+
+    @Override
+    public void writeFileContents(OutputStream out) throws IOException {
+        DatadogGlobalConfiguration globalConfiguration = DatadogUtilities.getDatadogGlobalDescriptor();
+        XSTREAM.toXMLUTF8(globalConfiguration, out);
+    }
+}