+++ src/main/java/org/datadog/jenkins/plugins/datadog/flare/ThreadDumpFlare.java
@@ -1,0 +1,48 @@
+package org.datadog.jenkins.plugins.datadog.flare;
+
+import hudson.Extension;
+
+import java.io.OutputStream;
+import java.io.PrintWriter;
+import java.nio.charset.StandardCharsets;
+import java.util.Map;
+
+@Extension
+public class ThreadDumpFlare implements FlareContributor {
+
+    @Override
+    public int order() {
+        return 7;
+    }
+
+    @Override
+    public String getDescription() {
+        return "Thread stack traces";
+    }
+
+    @Override
+    public String getFilename() {
+        return "thread-dump.txt";
+    }
+
+    @Override
+    public void writeFileContents(OutputStream out) {
+        // Print writer is not closed intentionally, to avoid closing out.
+        // Auto-flush set to true ensures everything is witten
+        PrintWriter printWriter = new PrintWriter(out, false, StandardCharsets.UTF_8);
+
+        Map<Thread, StackTraceElement[]> allStackTraces = Thread.getAllStackTraces();
+        for (Map.Entry<Thread, StackTraceElement[]> entry : allStackTraces.entrySet()) {
+            Thread thread = entry.getKey();
+            StackTraceElement[] stackTrace = entry.getValue();
+
+            printWriter.println("Thread: " + thread.getName());
+            for (StackTraceElement element : stackTrace) {
+                printWriter.println("    at " + element);
+            }
+            printWriter.println();
+        }
+
+        printWriter.flush();
+    }
+}